add frontend to enhance the app
Add the needed routes in the backend to have the full CRUD. Add routes for the answers. Render all questions and answers of the users. Add a login system to the routes and make it work in the frontend as well. Play with three.js and create a loading page.